backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
tree-sitter==0.23.2
python-dotenv==1.0.1
pydantic==2.9.2
